Convert solid.js to use Node require() / Browserify
Closes issue #8.
- Move library components to `./lib`
- Use require()
- Convert to using browserify / npm build script
- Update README
- Bump version to 0.5.0
