+++ VERSION
@@ -1,1 +1,1 @@
-0.4.0
+0.5.0
