+++ lib/meta.js
@@ -1,0 +1,6 @@
+'use strict'
+var lib = require('../package')
+
+module.exports.version = function version () {
+  return lib.version
+}
