+++ lib/web.js
@@ -1,52 +1,60 @@
-// LDP operations
-var Solid = Solid || {}
-Solid.web = (function () {
-  'use strict'
+'use strict'
+var XMLHttpRequest = require('xhr2')
+var parseLinkHeader = require('./utils').parseLinkHeader
 
-  // Init some defaults
-  var PROXY = 'https://databox.me/,proxy?uri={uri}'
-  var TIMEOUT = 5000
+// Init some defaults
+var PROXY = 'https://databox.me/,proxy?uri={uri}'
+var TIMEOUT = 5000
 
-  $rdf.Fetcher.crossSiteProxyTemplate = PROXY
-  // common vocabs
-  var LDP = $rdf.Namespace('http://www.w3.org/ns/ldp#')
+$rdf.Fetcher.crossSiteProxyTemplate = PROXY
+// common vocabs
+var LDP = $rdf.Namespace('http://www.w3.org/ns/ldp#')
 
-  // return metadata for a given request
-  function parseResponseMeta (resp) {
-    var h = Solid.utils.parseLinkHeader(resp.getResponseHeader('Link'))
-    var meta = {}
-    meta.url = (resp.getResponseHeader('Location')) ? resp.getResponseHeader('Location') : resp.responseURL
-    meta.acl = h['acl']
-    meta.meta = (h['meta']) ? h['meta'] : h['describedBy']
-    meta.user = (resp.getResponseHeader('User')) ? resp.getResponseHeader('User') : ''
-    meta.websocket = (resp.getResponseHeader('Updates-Via')) ? resp.getResponseHeader('Updates-Via') : ''
-    // writable/editable resource
-    meta.editable = []
-    var patch = resp.getResponseHeader('Accept-Patch')
-    if (patch && patch.indexOf('application/sparql-update') >= 0) {
-      meta.editable.push('patch')
+// return metadata for a given request
+function parseResponseMeta (resp) {
+  var h = parseLinkHeader(resp.getResponseHeader('Link'))
+  var meta = {}
+  meta.url = (resp.getResponseHeader('Location'))
+    ? resp.getResponseHeader('Location')
+    : resp.responseURL
+  meta.acl = h['acl']
+  meta.meta = (h['meta']) ? h['meta'] : h['describedBy']
+  meta.user = (resp.getResponseHeader('User'))
+    ? resp.getResponseHeader('User')
+    : ''
+  meta.websocket = (resp.getResponseHeader('Updates-Via'))
+    ? resp.getResponseHeader('Updates-Via')
+    : ''
+  // writable/editable resource
+  meta.editable = []
+  var patch = resp.getResponseHeader('Accept-Patch')
+  if (patch && patch.indexOf('application/sparql-update') >= 0) {
+    meta.editable.push('patch')
+  }
+  var allow = resp.getResponseHeader('Allow')
+  if (allow) {
+    if (allow.indexOf('PUT') >= 0) {
+      meta.editable.push('put')
     }
-    var allow = resp.getResponseHeader('Allow')
-    if (allow) {
-      if (allow.indexOf('PUT') >= 0) {
-        meta.editable.push('put')
-      }
-      if (allow.indexOf('POST') >= 0) {
-        meta.editable.push('post')
-      }
-      if (allow.indexOf('DELETE') >= 0) {
-        meta.editable.push('delete')
-      }
+    if (allow.indexOf('POST') >= 0) {
+      meta.editable.push('post')
     }
-
-    meta.exists = (resp.status === 200)
-    meta.xhr = resp
-    return meta
+    if (allow.indexOf('DELETE') >= 0) {
+      meta.editable.push('delete')
+    }
   }
 
-  // check if a resource exists and return useful Solid info (acl, meta, type, etc)
+  meta.exists = (resp.status === 200)
+  meta.xhr = resp
+  return meta
+}
+
+// LDP operations
+var client = {
+  // check if a resource exists and return useful Solid info
+  // (acl, meta, type, etc)
   // resolve(metaObj)
-  function head (url) {
+  head: function head (url) {
     var promise = new Promise(function (resolve) {
       var http = new XMLHttpRequest()
       http.open('HEAD', url)
@@ -60,16 +68,18 @@ Solid.web = (function () {
     })
 
     return promise
-  }
+  },
 
   // fetch an RDF resource
   // resolve(graph) | reject(this)
-  function get (url) {
+  get: function get (url) {
     var promise = new Promise(function (resolve, reject) {
       var g = $rdf.graph()
       var f = new $rdf.Fetcher(g, TIMEOUT)
 
-      var docURI = (url.indexOf('#') >= 0) ? url.slice(0, url.indexOf('#')) : url
+      var docURI = (url.indexOf('#') >= 0)
+        ? url.slice(0, url.indexOf('#'))
+        : url
       f.nowOrWhenFetched(docURI, undefined, function (ok, body, xhr) {
         if (!ok) {
           reject({status: xhr.status, xhr: xhr})
@@ -80,11 +90,11 @@ Solid.web = (function () {
     })
 
     return promise
-  }
+  },
 
   // create new resource
   // resolve(metaObj) | reject
-  function post (url, data, slug, isContainer, mime) {
+  post: function post (url, data, slug, isContainer, mime) {
     var resType = LDP('Resource').uri
     if (isContainer) {
       resType = LDP('BasicContainer').uri
@@ -117,11 +127,11 @@ Solid.web = (function () {
     })
 
     return promise
-  }
+  },
 
   // update/create resource using HTTP PUT
   // resolve(metaObj) | reject
-  function put (url, data, mime) {
+  put: function put (url, data, mime) {
     var promise = new Promise(function (resolve, reject) {
       mime = (mime) ? mime : 'text/turtle'
       var http = new XMLHttpRequest()
@@ -145,12 +155,12 @@ Solid.web = (function () {
     })
 
     return promise
-  }
+  },
 
   // patch a resource
   // accepts arrays of individual statements (turtle) as params
   // e.g. [ '<a> <b> <c> .', '<d> <e> <f> .']
-  function patch (url, toDel, toIns) {
+  patch: function patch (url, toDel, toIns) {
     var promise = new Promise(function (resolve, reject) {
       var data = ''
       var i
@@ -193,11 +203,11 @@ Solid.web = (function () {
     })
 
     return promise
-  }
+  },
 
   // delete a resource
   // resolve(true) | reject
-  function del (url) {
+  del: function del (url) {
     var promise = new Promise(function (resolve, reject) {
       var http = new XMLHttpRequest()
       http.open('DELETE', url)
@@ -216,18 +226,11 @@ Solid.web = (function () {
 
     return promise
   }
+}
 
-  // return public methods
-  // aliasing post -> create, put -> replace, patch -> update
-  return {
-    head: head,
-    get: get,
-    post: post,
-    create: post,
-    put: put,
-    replace: put,
-    patch: patch,
-    update: patch,
-    del: del
-  }
-}(this))
+// Alias some methods
+client.create = client.post
+client.replace = client.put
+client.update = client.patch
+
+module.exports = client
