+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "solid.js",
-  "version": "0.0.1",
+  "version": "0.5.0",
   "description": "Common library for writing Solid applications",
   "main": "dist/solid.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "browserify -r ./index.js:solid.js > dist/solid.js"
   },
   "repository": {
     "type": "git",
@@ -26,8 +27,14 @@
   },
   "homepage": "https://github.com/solid/solid.js",
   "dependencies": {
-    "rdflib": "^0.2.9"
+    "rdflib": "^0.2.10",
+    "xhr2": "*"
   },
   "devDependencies": {
+    "browserify": "*",
+    "standard": "^5.4.1"
+  },
+  "standard": {
+    "globals": [ "$rdf", "tabulator" ]
   }
 }
