--- status.js
@@ -1,26 +1,0 @@
-// Events
-var Solid = Solid || {}
-Solid.status = (function () {
-  'use strict'
-
-  // Get current online status
-  function isOnline () {
-    return window.navigator.onLine
-  }
-
-  // Is offline
-  function onOffline (callback) {
-    window.addEventListener('offline', callback, false)
-  }
-  // Is online
-  function onOnline (callback) {
-    window.addEventListener('online', callback, false)
-  }
-
-  // return public methods
-  return {
-    isOnline: isOnline,
-    onOffline: onOffline,
-    onOnline: onOnline
-  }
-}(this))
